fedeep/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fedeep"
version = "1.0.0"
description = "FedEEP — Federated Early-Exit with Progressive Phases (pure PyTorch)"
license = "Apache-2.0"
authors = [{ name = "Research Team" }]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=6.0",
    "medmnist>=3.0.0",
]

[project.optional-dependencies]
viz = ["tensorboard>=2.12.0", "matplotlib>=3.7.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
